feature(transfer-learning): added transfer learning experiment
This experiments on the idea of transfer learning, taking one of the two datasets, performing initial training, and then fine-tuning it on the other dataset. This is tested with each of the three models we have been working with (RandomForest, GradientBoosting, and XGBoost).

Also updated the requirements.txt for good measure since it was empty before. Should be able to download the python requirements easy now with -> pip install -r requirement.txt